setup: `IntentReveiver.java`: optimization
